+++ sdk/test-utils/recorder-new/test/testProxyTests.spec.ts
@@ -1,40 +1,10 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import {
-  createPipelineRequest,
-  HttpMethods,
-  PipelineRequestOptions
-} from "@azure/core-rest-pipeline";
 import { ServiceClient } from "@azure/core-client";
-import { env, isLiveMode, isPlaybackMode, Recorder } from "../src";
-import { expect } from "chai";
+import { isLiveMode, isPlaybackMode, Recorder } from "../src";
 import { TestMode } from "../src/utils/utils";
-
-const setTestMode = (mode: TestMode): TestMode => {
-  env.TEST_MODE = mode;
-  console.log(`==== setting TEST_MODE = ${mode} ====`);
-  return mode;
-};
-
-/**
- * Returns the test server url
- * Acts as the endpoint [ Works as a substitute to the actual Azure Services ]
- */
-function getTestServerUrl() {
-  // utils/server.ts creates a localhost server at port 8080
-  // - In "live" mode, we are hitting directly the localhost endpoint
-  // - In "record" and "playback" modes, we need to hit the localhost of the host network
-  //   from the proxy tool running in the docker container.
-  //   `host.docker.internal` alias can be used in the docker container to access host's network(localhost)
-  //
-  // if PROXY_MANUAL_START=true, we start the proxy tool using the dotnet tool instead of the `docker run` command
-  //  - in this case, we don't need to hit the localhost using the alias
-  //  - needed for the CI since we have difficulties with the mac machines
-  return !isLiveMode() && !(env.PROXY_MANUAL_START === "true")
-    ? `http://host.docker.internal:8080` // Accessing host's network(localhost) through docker container
-    : `http://127.0.0.1:8080`;
-}
+import { getTestServerUrl, makeRequestAndVerifyResponse, setTestMode } from "./utils/utils";
 
 // These tests require the following to be running in parallel
 // - utils/server.ts (to serve requests to act as a service)
@@ -44,14 +14,11 @@ function getTestServerUrl() {
     let recorder: Recorder;
     let client: ServiceClient;
 
-    const basePipelineReqOptions: Partial<PipelineRequestOptions> =
-      mode === "live" ? { allowInsecureConnection: true } : {};
-
     before(() => {
       setTestMode(mode);
     });
 
-    beforeEach(async function() {
+    beforeEach(async function () {
       recorder = new Recorder(this.currentTest);
       client = new ServiceClient({ baseUri: getTestServerUrl() });
       recorder.configureClient(client);
@@ -60,41 +27,10 @@ function getTestServerUrl() {
     afterEach(async () => {
       await recorder.stop();
     });
-
-    async function makeRequestAndVerifyResponse(
-      request: {
-        url?: string;
-        path: string;
-        body?: string;
-        headers?: { headerName: string; value: string }[];
-        method: HttpMethods;
-      },
-      expectedResponse: { [key: string]: unknown } | undefined
-    ) {
-      const req = createPipelineRequest({
-        url: request.url ?? getTestServerUrl() + request.path,
-        body: request.body,
-        method: request.method,
-        ...basePipelineReqOptions
-      });
-      request.headers?.forEach(({ headerName, value }) => {
-        req.headers.set(headerName, value);
-      });
-      const response = await client.sendRequest(req);
-      if (expectedResponse) {
-        if (!response.bodyAsText) {
-          throw new Error("Expected response.bodyAsText to be defined");
-        }
-
-        expect(JSON.parse(response.bodyAsText)).to.deep.equal(expectedResponse);
-      }
-      // Add code to also check expected headers
-      return response;
-    }
-
+    
     it("sample_response", async () => {
       await recorder.start({ envSetupForPlayback: {} });
-      await makeRequestAndVerifyResponse(
+      await makeRequestAndVerifyResponse(client,
         { path: `/sample_response`, method: "GET" },
         { val: "abc" }
       );
@@ -103,7 +39,7 @@ function getTestServerUrl() {
     it("sample_response with random string in path", async () => {
       await recorder.start({ envSetupForPlayback: {} });
 
-      await makeRequestAndVerifyResponse(
+      await makeRequestAndVerifyResponse(client,
         {
           path: `/sample_response/${recorder.variable(
             "random-1",
@@ -113,7 +49,7 @@ function getTestServerUrl() {
         },
         { val: "I am the answer!" }
       );
-      await makeRequestAndVerifyResponse(
+      await makeRequestAndVerifyResponse(client,
         {
           path: `/sample_response/${recorder.variable("random-2", "known-string")}`,
           method: "GET"
@@ -122,275 +58,7 @@ function getTestServerUrl() {
       );
     });
 
-    describe("Sanitizers", () => {
-      it("GeneralRegexSanitizer", async () => {
-        env.SECRET_INFO = "abcdef";
-        const fakeSecretInfo = "fake_secret_info";
-        await recorder.start({
-          envSetupForPlayback: {
-            SECRET_INFO: fakeSecretInfo
-          }
-        }); // Adds generalRegexSanitizers by default based on envSetupForPlayback
-        await makeRequestAndVerifyResponse(
-          {
-            path: `/sample_response/${env.SECRET_INFO}`,
-            method: "GET"
-          },
-          { val: "I am the answer!" }
-        );
-      });
-
-      it("RemoveHeaderSanitizer", async () => {
-        await recorder.start({
-          envSetupForPlayback: {},
-          sanitizerOptions: {
-            removeHeaderSanitizer: {
-              headersForRemoval: ["ETag", "Date"]
-            }
-          }
-        });
-        await makeRequestAndVerifyResponse(
-          { path: `/sample_response`, method: "GET" },
-          { val: "abc" }
-        );
-      });
-
-      it("BodyKeySanitizer", async () => {
-        const secretValue = "ab12cd34ef";
-        const fakeSecretValue = "fake_secret_info";
-        await recorder.start({
-          envSetupForPlayback: {},
-          sanitizerOptions: {
-            bodyKeySanitizers: [
-              {
-                jsonPath: "$.secret_info", // Handles the request body
-                regex: secretValue,
-                value: fakeSecretValue
-              },
-              {
-                jsonPath: "$.bodyProvided.secret_info", // Handles the response body
-                regex: secretValue,
-                value: fakeSecretValue
-              }
-            ]
-          }
-        });
-        const reqBody = {
-          secret_info: isPlaybackMode() ? fakeSecretValue : secretValue
-        };
-        await makeRequestAndVerifyResponse(
-          {
-            path: `/api/sample_request_body`,
-            body: JSON.stringify(reqBody),
-            method: "POST",
-            headers: [{ headerName: "Content-Type", value: "application/json" }]
-          },
-          { bodyProvided: reqBody }
-        );
-      });
-
-      it("BodyRegexSanitizer", async () => {
-        const secretValue = "ab12cd34ef";
-        const fakeSecretValue = "fake_secret_info";
-        await recorder.start({
-          envSetupForPlayback: {},
-          sanitizerOptions: {
-            bodyRegexSanitizers: [
-              {
-                regex: "(.*)&SECRET=(?<secret_content>[^&]*)&(.*)",
-                value: fakeSecretValue,
-                groupForReplace: "secret_content"
-              }
-            ]
-          }
-        });
-        const reqBody = `non_secret=i'm_no_secret&SECRET=${
-          isPlaybackMode() ? fakeSecretValue : secretValue
-        }&random=random`;
-        await makeRequestAndVerifyResponse(
-          {
-            path: `/api/sample_request_body`,
-            body: reqBody,
-            method: "POST",
-            headers: [{ headerName: "Content-Type", value: "text/plain" }]
-          },
-          { bodyProvided: reqBody }
-        );
-      });
-
-      it("UriRegexSanitizer", async () => {
-        const secretEndpoint = "host.docker.internal";
-        const fakeEndpoint = "fake_endpoint";
-        await recorder.start({
-          envSetupForPlayback: {},
-          sanitizerOptions: {
-            uriRegexSanitizers: [
-              {
-                regex: secretEndpoint,
-                value: fakeEndpoint
-              }
-            ]
-          }
-        });
-        const pathToHit = `/api/sample_request_body`;
-        await makeRequestAndVerifyResponse(
-          {
-            url: isPlaybackMode()
-              ? getTestServerUrl().replace(secretEndpoint, fakeEndpoint) + pathToHit
-              : undefined,
-            path: pathToHit,
-            method: "POST"
-          },
-          { bodyProvided: {} }
-        );
-      });
-
-      it("UriSubscriptionIdSanitizer", async () => {
-        const id = "73c83158-bd73-4cda-aa11-a0c2a34e2544";
-        const fakeId = "00000000-0000-0000-0000-000000000000";
-        await recorder.start({
-          envSetupForPlayback: {},
-          sanitizerOptions: {
-            uriSubscriptionIdSanitizer: {
-              value: fakeId
-            }
-          }
-        });
-        await makeRequestAndVerifyResponse(
-          {
-            path: `/subscriptions/${isPlaybackMode() ? fakeId : id}`,
-            method: "GET"
-          },
-          { val: "I am the answer!" }
-        );
-      });
-
-      it("ContinuationSanitizer", async () => {
-        await recorder.start({
-          envSetupForPlayback: {},
-          sanitizerOptions: {
-            continuationSanitizers: [
-              {
-                key: "your_uuid",
-                method: "guid", // What is this method exactly?
-                resetAfterFirst: false
-              }
-            ]
-          }
-        });
-        // What if the id is part of the response body and not response headers?
-
-        const firstResponse = await makeRequestAndVerifyResponse(
-          {
-            path: `/api/sample_uuid_in_header`,
-            method: "GET"
-          },
-          undefined
-        );
-
-        await makeRequestAndVerifyResponse(
-          {
-            path: `/sample_response`,
-            method: "GET",
-            headers: [
-              {
-                headerName: "your_uuid",
-                value: firstResponse.headers.get("your_uuid") || ""
-              }
-            ]
-          },
-          { val: "abc" }
-        );
-      });
-
-      it("HeaderRegexSanitizer", async () => {
-        const sanitizedValue = "Sanitized";
-        await recorder.start({
-          envSetupForPlayback: {},
-          sanitizerOptions: {
-            headerRegexSanitizers: [
-              {
-                key: "your_uuid",
-                value: sanitizedValue
-              }
-            ]
-          }
-        });
-
-        await makeRequestAndVerifyResponse(
-          {
-            path: `/api/sample_uuid_in_header`,
-            method: "GET"
-          },
-          undefined
-        );
-        // TODO: Add more tests to cover groupForReplace
-      });
-
-      // it("OAuthResponseSanitizer", async () => {
-      //   await recorder.start({});
-      //   await recorder.addSanitizers({
-      //     oAuthResponseSanitizer: true
-      //   });
-
-      //   await makeRequestAndVerifyResponse(
-      //     {
-      //       path: `/api/sample_uuid_in_header`,
-      //       method: "GET"
-      //     },
-      //     undefined
-      //   );
-      //   // TODO: Add more tests to cover groupForReplace
-      // });
-
-      it.skip("ResetSanitizer (uses BodyRegexSanitizer as example)", async () => {
-        const secretValue = "ab12cd34ef";
-        const fakeSecretValue = "fake_secret_info";
-        await recorder.start({
-          envSetupForPlayback: {},
-          sanitizerOptions: {
-            bodyRegexSanitizers: [
-              {
-                regex: "(.*)&SECRET=(?<secret_content>[^&]*)&(.*)",
-                value: fakeSecretValue,
-                groupForReplace: "secret_content"
-              }
-            ]
-          }
-        });
-        const reqBody = `non_secret=i'm_no_secret&SECRET=${
-          isPlaybackMode() ? fakeSecretValue : secretValue
-        }&random=random`;
-        await makeRequestAndVerifyResponse(
-          {
-            path: `/api/sample_request_body`,
-            body: reqBody,
-            method: "POST",
-            headers: [{ headerName: "Content-Type", value: "text/plain" }]
-          },
-          { bodyProvided: reqBody }
-        );
-
-        await recorder.addSanitizers({
-          resetSanitizer: true
-        });
-
-        const reqBodyAfterReset = `non_secret=i'm_no_secret&SECRET=${secretValue}&random=random`;
-        // TODO: BUG OBSERVED - The following request should not be sanitized, but is sanitized
-        await makeRequestAndVerifyResponse(
-          {
-            path: `/api/sample_request_body`,
-            body: reqBodyAfterReset,
-            method: "POST",
-            headers: [{ headerName: "Content-Type", value: "text/plain" }]
-          },
-          { bodyProvided: reqBodyAfterReset }
-        );
-      });
-    });
-
     // Matchers
-
     describe("Matchers", () => {
       it("BodilessMatcher", async () => {
         await recorder.start({ envSetupForPlayback: {} });
@@ -400,7 +68,7 @@ function getTestServerUrl() {
         // different body in playback vs record mode.
         const body = isPlaybackMode() ? "playback" : "record";
 
-        await makeRequestAndVerifyResponse(
+        await makeRequestAndVerifyResponse(client,
           {
             path: `/sample_response`,
             body,
@@ -420,7 +88,7 @@ function getTestServerUrl() {
           value: isPlaybackMode() ? "playback" : "record"
         };
 
-        await makeRequestAndVerifyResponse(
+        await makeRequestAndVerifyResponse(client,
           {
             path: `/sample_response`,
             body: "body",
